+++ lib/commons/molecules/listfield/list_field_model.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+
+class ListFieldModel {
+  String id;
+  TextEditingController textEditingController;
+  bool isEdit = true;
+
+  ListFieldModel({this.id, this.isEdit, this.textEditingController});
+}