+++ CHANGELOG.md
@@ -73,3 +73,12 @@ Add font color for General DS Button
 ## [0.0.1+18] - 17 February 2021.
 
 Add General DS Ghost Button
+
+## [0.0.1+20] - 20 December 2021.
+
+Add General DS List Text Field
+
+## [0.0.1+21] - 20 December 2021.
+
+Fixing DS List Text Field
+